Adicionada a possibilidade de adicionar variáveis de ambiente por meio do arquivo set_env.json

+++ part-1/test/rest_api.test.js
@@ -1,3 +1,4 @@
+import loadEnvironmentVariablesIfAny from '../src/load_env.js'
 import axios from "axios";
 import mocha from 'mocha';
 import chai  from 'chai'
@@ -7,12 +8,17 @@ const it = mocha.it;
 const expect = chai.expect;
 const assert = chai.assert;
 
+loadEnvironmentVariablesIfAny();
+
+const environment = process.env.NODEJS_APP_ENVIRONMENT === "production" ? process.env.NODEJS_APP_ENVIRONMENT : "development";
+
 describe("Teste de requisições para a API desenvolvida", function(){
-    const apiHost = process.env.API_HOST || "localhost";
-    const apiPort = process.env.API_PORT || "3001";
+    const apiHost = process.env.NODEJS_API_HOST || "localhost";
+    const apiPort = process.env.NODEJS_API_PORT || "3001";
     const apiBaseUrl = `http://${apiHost}:${apiPort}`;
 
     let initialRequestSuccessed = false;
+    let resetSuccessed = false;
     let requestOptions = {
         maxContentLength: Infinity,
         maxBodyLength: Infinity
@@ -34,25 +40,48 @@ describe("Teste de requisições para a API desenvolvida", function(){
         }
     });
 
-    it("Requisição para resetar a API deve retornar status 200", async function(){
+    it("Requisição para resetar a API deve retornar status 200 caso o ambiente não seja de produção", async function(){
+        if(!initialRequestSuccessed){
+            assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
+        }
+        if(environment === "development") {
+            try {
+                let res = await axios.delete(apiBaseUrl + "/reset",
+                    requestDataForTest,
+                    requestOptions);
+
+                expect(res.status).to.equal(200);
+                resetSuccessed = true;
+            } catch (err) {
+                assert.fail(err.message);
+            }
+        }
+    });
+
+    it("Requisição para resetar a API deve retornar status 403 caso o ambiente seja de produção", async function(){
         if(!initialRequestSuccessed){
             assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
         }
 
-        try{
-            let res = await axios.delete(apiBaseUrl+"/reset",
-                requestDataForTest,
-                requestOptions);
+        if(environment === "production") {
+            try {
+                let res = await axios.delete(apiBaseUrl + "/reset",
+                    requestDataForTest,
+                    requestOptions);
 
-            expect(res.status).to.equal(200);
-        }catch(err){
-            assert.fail(err.message);
+                assert.fail("A requisição deveria ter retornado status 403!");
+            } catch (err) {
+                expect(err.response.status).to.equal(403);
+            }
         }
     });
 
     it("O envio da primeira requisição POST deve retornar status 200", async function(){
         if(!initialRequestSuccessed){
             assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
+        }
+        if(!resetSuccessed){
+            assert.fail("Condição inicial do teste não foi cumprida! Houve um erro no reset da API!");
         }
 
         try{
@@ -69,6 +98,9 @@ describe("Teste de requisições para a API desenvolvida", function(){
     it("As próximas 100 requisições com o mesmo corpo do item anterior devem retornar status 403", async function(){
         if(!initialRequestSuccessed){
             assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
+        }
+        if(!resetSuccessed){
+            assert.fail("Condição inicial do teste não foi cumprida! Houve um erro no reset da API!");
         }
 
         for(let i=0; i < 100; i++){
@@ -87,6 +119,9 @@ describe("Teste de requisições para a API desenvolvida", function(){
         if(!initialRequestSuccessed){
             assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
         }
+        if(!resetSuccessed){
+            assert.fail("Condição inicial do teste não foi cumprida! Houve um erro no reset da API!");
+        }
 
         requestDataForTest.text += "a";
 
@@ -104,6 +139,9 @@ describe("Teste de requisições para a API desenvolvida", function(){
     it("O envio de uma requisição POST com um corpo de tamanho aproximado de 1GB deve retornar status 200", async function(){
         if(!initialRequestSuccessed){
             assert.fail("Condição inicial do teste não foi cumprida! Houve um erro na requisição inicial!");
+        }
+        if(!resetSuccessed){
+            assert.fail("Condição inicial do teste não foi cumprida! Houve um erro no reset da API!");
         }
 
         let bigObject;
